pyproject.toml
[project]
name = "agentic-slackbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.6.0",
    "openai-agents>=0.0.13",
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
]
